Split routes into dedicated pages
